New much improved AutoWidget class.
Instead of a using a separate check button to switch from automatic
to manual mode, the new AutoWidget is just a single Gtk.Entry.  If
it's empty, it displays "automatic" in gray text.  If the user types
anything, "automatic" is deleted and replaced by the user's text
in black.

GenericWidgets now can take an optional 'value' constructor argument,
which is used instead of param.value when initializing the widget.
This gets around the problem that AutoWidget may not be able to
query param.value at widget construction time, because the parameter
value may need to be computed from other widgets that aren't
constructed yet.  This wasn't a problem with the old AutoWidget
because it wasn't derived from GenericWidget.

This didn't really have to be done as part of the gtk3 upgrade, but
this is a reasonable time to do it, since it is a similar sort of
UI change from the user's perspective.  Plus, I felt like it.