Dialog boxes now appear over the windows that created them.
I had built the previous commit on Linux, and there gtk3 was
complaining that some dialogs didn't have a "transient parent".
Now our SubWindow and ParameterDialog classes require a parentwindow
argument, as do functions like getParameters() which use ParameterDialog.
(Most of the changes in this commit are just adding that argument
to getParameters() calls.) This allows the subwindows to appear
over the window that was being used when the subwindow was created,
instead of appearing at some other random point on the screen,
possibly behind something else.  On Mac at least the subwindow can
be moved independently of the original window, but if the original
window is moved, the subwindow goes with it.

Since the Quit command can be run from many different windows I had
to resort to some ugliness to give it the correct transient parent
in all cases, but there are comments.

Some windows derived from SubWindow weren't using the SubWindow's
File menu, for no good reason (that I know of, yet).  I fixed some
of them, but I broke the ActivityViewer.

The thing that I thought I was investigating when I built the last
commit on Linux was the new AutoWidget.  It works, except that there
are more gtk themes available on Linux, and the assumptions I was
making about the text color were incorrect, and the widgets can be
unreadable.  It will be necessary to use gtk CSS properly, I guess.
This commit contains a bit of exploratory code.