An AutoWidget in automatic mode now displays its text in italics.
This should be enough to distinguish the automatic identifier
from text entered by the user.

I learned the correct way to change the widget's style, using a Gtk
CSS class.  The previous version changed the text color using a
deprecated interface.  This version doesn't change the text or
background color, so that it doesn't have to worry about what colors
the Gtk theme is using.  (But what if the theme already uses italics?)
There are commments about how to change the text color if that turns
out to be useful.

Moved the global addStyle function from oofGUI.py to gtkutils.py,
which makes more sense anyway.  parameterwidgets.py couldn't import
it from oofGUI.py because of an import loop.