Now using Gtk.Stacks in the main OOF window and in the Chooser widgets.
Gtk.Stack is a better way of switching between widgets.  Using it
to switch pages allows the main window size to stay fixed.  Using
it in the Chooser widget allows the page switching widget at the
top of the main OOF GUI window to remain a fixed size, which means
that you can park your mouse on the next page or prev page buttons
and click multiple times, without the button moving out from under
the mouse.

Next is to use a Gtk.Stack for switching toolboxes in the graphics
window.