Simplifications in the gfxwindow calling sequence, in particular for hiding and showing layers.
The contour map was being drawn twice when the contourable layer
changed, which turned out to be symptom of a lot of redundant calls.
Some of the redundancy is in the pre-gtk3 version, too, but I'm not
going to track it down.

There's still a problem with the contourmap, which isn't disappearing
immediately when the contourable layer is hidden or deleted.  It
disappears when something generates another "draw" event, like
moving the mouse into the window.

Deleted a TODO about fixing the pop-up menu in the gfxwindow layer
list.  It doesn't need to be fixed. I just hadn't enable right-clicks
on the mouse.