Chooser.update() now indicates if anything changed.
A gtklogger checkpoint in the Analysis Page was being reached far
too often because a Chooser was being updated too often.  Now the
Chooser widget indicates if it has actually changed, so the rest
of the program can react appropriately.  I left in a comment to try
to find out why the chooser was being updated too often.