GUI logging of events on the OOFCanvas is working.
This was much easier than I expected it to be, since the
existing gtklogger code mostly just worked on the OOFCanvas's
GtkLayout.  (It was difficult in the old gtk+2 code because
the libgnomecanvas object wasn't a PyGTK object.)

The signal and event callbacks in the canvas are now careful
to return a bool, which controls whether or not further handlers
are invoked (such as the gtklogger handlers).

Logging of motion events is now turned on and off by changing
the return value of the OOFCanvas's motion event handler.  This
is much simpler than using the gtklogger's log_motion_events()
method.

Removed some obsolete code.  Added some comments and TODOs.