Canvas mouse events are now logged using user coordinates instead of device coordinates.
This is supposed to make the log files portable.  The trouble with
logging events using device coordinates is that the conversion from
device to user coordinates depends on the canvas's widget size, and
that can change with screen resolution or Gtk theme or widget font
size.  Since all the real work happens in user coordinates, logging
the user coordinate instead of the device coordinate should avoid
the problem.

Because converting to user coordinates can't be done by the GtkLayout,
logging has to be done by the Canvas, and we need a mechanism for
associating a particular logger with a particular non-Gtk object.
So now all of the gtklogger.connect* methods can take an optional
"logger" keyword argument which tells them to use the given object
instead of creating a logger in the usual way.

The Canvas now has a swigged pixel2user function that allows the
coordinate transformation to be done from Python.