Mostly comments and minor changes.
Added a --verbose-switchboard start up option, in order to track
down a bug that was cropping up during GUI construction.

Increased the number of digits record when logging mouse events on
the canvas.  This increases the chances that a OOF script saved
after recording a gui log file will agree with the script saved
after replaying the gui log file.