Fixes to get the AutoWidget to work with gtklogger.
gtklogger relied on the 'insert_text' signal when logging text
insertion events in Gtk.Entries.  It turns out that there's a bug
in Gtk3.  Signal handlers that are supposed to return a value to
Python callers can't do it, and if the value is in an in/out parameter
in C, the input value is wrong as well.  The solution is to use a
subclass of Gtk.Entry and override the subclass method that in a
better world would emit the signal.

I've added an AutomaticEntry class that is derived from Gtk.Entry.
AutoWidget (no longer derived from GenericWidget) use an AutomaticEntry
instead of a Gtk.Entry.  Its do_insert_text() method handles
everything that the 'insert_text' handler would, both for doing the
actual text insertion and for gui logging.

Comments in the code refer to the stackoverflow pages where I
learned how to do this.