Fixed a couple of gtklogger problems.
Rerecording wasn't working because the flag for starting the
comment-adding gui wasn't supported.  Changed the name of the
flag from logger_comments to comment_gui.

The 'focus_out_event' logger now includes a check that the widget
still exists.  A focus_out_event can be caused by widget destruction.
When replaying, the event that cause the widget to be destroyed will
be replayed, and then gtklogger will try to make the widget emit its
focus_out_event signal, which is both impossible and redundant at
that point.