Replaying is now done via Gtk.main_do_event() rather than Widget.event().
This is supposed to be better, according to the docs.

The gui scripts are more compact, because there's mostly no need
to set a variable containing a widget path.  Also, the event()
function now longer just constructs an event.  It also passes it
to Gtk.main_do_event(), reducing the amount of code that has to
be explicit in the log file.

WindowLogger had some complicated code related to delete events,
which doesn't seem to be either necessary or possible when using
Gtk.main_do_event(), so it's now simpler too.