Updated the first two gui test log files so that they use the new event handling, and added tests to them.
Changed the way that generics.chooserCheck works.  It can't check
that the choices in the Chooser are in the right order, because
that order isn't known by the Gtk part of the widget, which is
the part that is accessible to gtklogger.  The GtkStack on which
the Chooser is based uses a dictionary and not a list, apparently.

materialmanaager.getMaterialNames() now returns a sorted list, so
materials listed in the GUI should always be in alphabetical order.