Changed autogroup so that it returns a list of new group names, instead of just one.
Autogroup creates pixel groups automatically.  It used to return
the name of the last group that it created, so that the new group
could be selected in the GUI.  However, it used a std::map to
store the groups, and so the order in which they were created was
system dependent, and GUI tests weren't portable.   Now it returns
a list of group names from C++ to python, and python sorts the
list and selects the first one.

I'm not sure why this wasn't a problem with the gtk+2 version.
Probably we were just lucky.