The tutorial window no longer puts its text in a scrolled window.
Scrolling the text was a problem because it didn't work if a
modal dialog was open, so in a few places in a tutorial had to
say "make sure the window is big enough to read the rest of this.."
which was dumb.  Now the window is always big enough.  Tutorial steps
should always be short enough to fit in a reasonably sized window.

Minor changes to the simple example tutorial, including removing
the instruction to make sure the window is big enough.