Rerecorded the simple example tutorial test with the current mods to the ChooserWidget.
The ChooserWidget logging problems aren't solved completely yet.
When an item is selected in the widget, the gui log contains both
a 'deactivate' event for the menu and an 'activate' event for the
menu item.  The 'deactivate' is redundant because the 'activate'
also closes the menu.  This sometimes but not always raises a
Gdk-CRITICAL exception, so there must be a race condition of some
sort.  On the other hand, the program appears to behave correctly
even with the error.