loggergui now does post-processing, as well as adding comments to log files.
The minimal postprocessing that had been done in loggers.py is
now done in loggergui, which runs in a separate process.  It can
now remove the *previous* line from a log file if the next line
is redundant with it.  The most important situation concerns
menu deactivation and menu item activation, especially for the
ChooserWidget.