Debugging a gtk problem.
When switching to the Materials page, the first entry in the Property
list is automatically and magically selected if there's a Microstructure
and if the page is switched with the Chooser.  It doesn't happen
if no Microstructure has been created, or if the arrow buttons are
used to switch pages.