New branch to debug the peculiar interaction between the main page Chooser and the Property Tree widget.
The pathology is that creating a Microstructure and then switching
to the Materials page using the pull-down page chooser selects
the first entry in the Property tree.

chooser.py now contains an OLDChooserWidget, which is the gtk2 one
updated for gtk3 but without tooltips (the ComboBox tooltip hack
doesn't work in gtk3) and a NEWChooserWidget, which is the new hack,
using a GtkEventBox, GtkStack, and GtkMenu.

Using the OLDChooserWidget, the problem goes away (but there are no
tooltips).

All pages except the Microstructure and Materials pages are commented out.