Fixed the bug that was sometimes causing odd behavior when switching pages in the main GUI.
When using a GtkStack to switch pages, if an element of a page had
focus when the page was removed, an element of the replacement page
was given focus when it was installed.  This led to surprises in
the GUI tests.  The problem didn't appear when using the old,
non-GtkStack scheme, because then the old pages were actually removed
from the GUI and not just hidden.  (Presumably the GtkStack has to
ensure that the focus is transferred so that hidden objects don't
receive events.)  It also didn't appear when using the old GtkComboBox
based ChooserWidget, because it grabbed focus when it was used.

The new ChooserWidget for switching pages now grabs focus when it's
used, but it doesn't actually use it, in that it can't respond
keypress events.  I'm investigating how to make it do that, but the
code is commented out for now.