Improvements to the ChooserWidget focus handling.
The widget's GtkStack now grabs the focus after the menu closes,
which makes it unnecessary for the oofGUI page to force it to grab
focus.

Hitting <Return> when the widget has focus now brings up the menu,
which can be navigated with the arrow keys.  The arrow keys still
have to somehow be told which menu item is initially selected,
though.

There still needs to be some visual feedback that the ChooserWidget
has focus.

Removed the debugging code that made using the GtkStack optional
in the main window.