Tweaks to the ChooserWidget.
Although the docs recommend giving EventBoxes invisible windows and
keeping them above their children, that seems to interfere with the
gui logging, so it's commented out.

The current entry in the ChooserWidget is now desensitized and in
italics in the pop-up menu.  This is a poor substitute for the
GtkComboBox behavior of presenting the menu with the current item
selected, but the GtkComboBox doesn't allow tooltips, and we want
tooltips.   At least this is better than the previous attempt, which
used GtkRadioMenuItems and was horribly ugly.