Fixed bugs in RegisteredClassFactory and ParameterTable, and added a feature to ChooserWidget.
RegisteredClassFactory had been broken by an earlier change that
was made to maintain keyboard focus after the widget was updated.
That change made the widget not update at all when set from a
Registration, if the current Registration class hadn't changed.

In gtk3 ParameterTable apparently needs to explicitly delete its old
widgets when it's updated.

ChooserWidget now has an optional allowNone argument for __init__.
If allowNone is True, then the widget can be set to None, even if
it has a non-empty list of choices.  It displays "---" if the value
is None.  This will be used in the namedAnalysisChooser on the
AnalyzePage.