Fixes to the Named Analysis parts of the Analysis and Boundary Analysis pages.
Changes were required by the gtk3 version of the ChooserWidget, and
take advantage of the new widget's "allowNone" flag.  The widget
has no value if the current settings of the Output/Domain/etc
widgets don't correspond to a named analysis.

Also removed the pop-up menu for the named analysis operations and
replaced it with three buttons. There's room for the buttons and it's
more consistent with the rest of the interface.

Code cleanup:

Replaced some IntParameters with PositiveIntParameters.

Changed the way that the signalling loop is suppressed for
the named analysis chooser widget.  Now it just blocks
some signals, instead of adding a new flag to check.

The findNamedXXXXAnalysis methods in namedanalysis.py now return
None instead of "" if they can't find anything, because that can
be passed directly to the ChooserWidget.

Deleted some old code.