Changed the way menu items are referred to in gtklogger, and re-recorded the gui log of the first tutorial.
Hierarchical menu items used to be specified in gui log files as a
colon separated string, like 'Menu:Item:SubItem'.  This was a bad
idea, because in principle the menu item names could contain colons.
Now they're saved as a list of strings, ['Menu','Item','SubItem'].

Fixed chooserCheck in generics.py so that it works with the modified
ChooserWidget.