Fixed (sort of) a race condition when double clicking on the graphics layer list.
Double clicking (activating) to edit an entry in the graphics
window's layer list invokes the oofmenu callback for selecting the
layer and then the callback for activating it.  If the second
callback runs before the first callback finishes, the wrong layer
can be loaded into the editing dialog box.  I implemented the simple
solution, which was to force the first callback to run on the main
thread.  This feels like the wrong scope, but it's straightforward.

I don't know why this wasn't a problem with gtk+2.  Perhaps the timing
of the callback has changed.