First pass at a ChunkyVector class.
ChunkyVector is like std::vector, but it allocates space in chunks
instead of maintaining a contiguous stretch of memory.  It might
be faster and avoid memory fragmentation in some cases because it
never reallocates anything except for an index of chunks.

This version doesn't allow entries to be removed from the vector.
It has push_back, emplace_back, operator[], and iterators.

Some of the intermediate data structures in LinearizedSystem are
now ChunkyVectors.