Added wevent() as an alternative to event() when recreating Gdk events in a log file.
The gtk documentation is vague on how to do this correctly, because
apparently it's not a good idea.  wevent() uses GtkWidget.event(),
which ensures that the event is associated with the correct widget.
However the docs say to use Gtk.main_do_event() instead, so that
the event will behave as if it's in the event queue.  However,
main_do_event() takes a window and a position as arguments, not a
widget, and sometimes the wrong widget responds when replaying a
log file.