Fixed a replay bug.
When a gui log file failed to replay because it gave up waiting for
a top level window to appear, the next log line in the queue was
invoked in a perpetual timeout callback.  This wasn't a big problem
because usually the next step after a replay failure is to quit
the program, but it was making debugging difficult.  Now the
replay mechanism quits properly in this situation.