Added debugOptions for investigating various pop-up menu fixes.  Also modifier key handling in gtklogger.
gfxmenu.py, menulogger.py, widgetlogger.py all have a
debugOption variable which can choose between different
code snippets for handing pop-up menus.  All of that will
be deleted in the next commit because I think that the thing
to do is to not log the pop-up menus at all.  I'd have
committed the changes on a branch but I want to keep the
modifier key changes (below).  This shows poor planning on
my part.

Modifier key press and release events should not be logged.  They
do strange things when replayed, and they're not needed as long
as the modifier key state is recorded with each key press and
mouse button event.