Changed BooleanParameter so that its value is always True or False, not 1 or 0.
Changes elsewhere so that the flags for modifier keys are also either
True or False, not 1 or 0.

There had a been a TODO about this.  It was actually a problem
because inconsistent treatment was making log files that should
have been identical different, and breaking tests.  For example,
the state of the shift and control keys was reported as 0 or 1 when
recording a mouse click, but True or False when repeating the click
with the "Repeat" button.  When replayed in a gui test, only 0 and
1 were reported, so the log files appeared to be different.

This change may break some test scripts, but since I'm still early
in the process of re-recording the GUI tests for Gtk3, it's a good
time to do it.  I don't think it will break any user OOF2 scripts.