Instrumented the skeleton tutorial gui test.
There aren't many tests in it, actually.  I just copied what
was in the gtk2 version.

The gtk2 version has a comment saying that it doesn't do
exactly what's in the tutorial, because some of it isn't
portable.  This may need to be re-recorded again if that's
still the case.