Fixed a bug that arose while deleting Microstructures.
Deleting a Who object is calling GhostGfxWindow.drawIfNecessary
after each nested Who object is deleted (which is too often, but
not the subject of this fix).  The ActiveArea layer was
being drawn after it was destroyed.  DisplayMethod.incomputable()
now checks that a proxy Who object can be resolved, and not just
that it's not None.  This fixes the problem.  I don't know why
the problem didn't exist before the gtk3 conversion.  Presumably
objects were being destroyed in a different order.

Added a TODO about not redrawing until all objects are destroyed.