Changed the way in which DefaultLayers work in graphics windows.
Graphics windows now have a "New Layer Policy" setting that
governs how default graphics layers are created.  The setting
can be changed in the graphics window's Settings menu, and
a default setting for new windows is in the main window's
Settings/Graphics Defaults menu.

This is meant to address the confusion that Edwin reported
when users don't see what they expect in the graphics window.

The possible settings are:

"Never": no default layers are created, and new gfxwindows are empty.
This is the default.

"Single": default layers are created in a new window for pre-existing
Images, Skeletons, and Meshes *if* there is only one possible object
to display.  New Images, Skeletons, and Meshes are displayed in
pre-existing windows if the window doesn't already contain one of
the objects.  This is sort of like the old behavior, except that
if you delete all layers for a particular type of object, the next
object of that type that's created will be displayed.  The old
version assumed that if you deleted layers you wanted them to stay
deleted.

"Always": All new Images, Skeletons, and Meshes will be displayed
automatically in pre-existing graphics windows.  New graphics windows
will be empty.

There could be another option that works like "Always" but also
displays all pre-existing objects in new graphics windows.