Fixed a bug in FloatRangeWidget and IntRangeWidget.
The LabelledSlider wasn't calling its callback when its GtkEntry
was empty or contained un-evaluatable text.   This was wrong because
it prevented things that used the LabelledSlider from sensitizing
themselves properly.  Now the callback is called and the calling
procedure is responsible for noticing if the widget is in a consistent
state.

Added LabelledSlider.consistent() to help determine if the widget
is in a consistent state.

The released gtk2 version also has this bug.