Added memory monitoring.  It only works in Linux so far.
OOF.Help.Debug.Memory_Monitor.Start(filename=...) turns it on, and
OOF.Help.Debug.Memory_Monitor.Stop() turns if off.  From inside the
code, call debug.memusage(comment) at points where you want to know how
much memory is being used.  It will print the usage in KB to the file
specified in the Start command, followed by the comment.

This uses pmap to get the memory usage.  It's reporting the "total"
from the last line of the pmap output.

I put a few representative calls to memusage in a few places in the
solution step for linear problems, but make no guarantees that they're
in the best spots.