Made memusage() available in oof scripts.
This hasn't been tested because I accidentally made the changes
on the Mac, where memusage() doesn't work.  Transferring the files
to Linux is easist via git.

memusage() is now in utils.py instead of debug.py to avoid an
import loop, and utils.py isn't a bad place for it anyway.