Updated the Move Nodes tutorial, and re-recorded and instrumented the associated gui test.
Added a comment in ghostgfxwindow.py about a lock on the list of
layers.  The pre-gtk3 version had a lock in the Display class that
was different from the gfxlock in the gfx window.  The Display lock
was an SLock so it could be used on the main thread. In particular,
it was used in GhostGfxWindow.drawable(), which the gui tests use
to determine whether or not the window has anything drawable in it.
In the current version, there is no Display class, and the only
lock is the window's gfx lock, which is taking the role of both
locks in the earlier version.  That means that drawable() can't use
it, because the gui testing code runs on the main thread.

Instead of adding an additional SLock on the layer list, I've just
removed the lock check from GhostGfxWindow.drawable().  Since it's
only used in the gui test, it's probably safe enough.

Updated some code in generics.py for gtk3.