Fixed a graphics window bug.
GhostGfxWindow.incorporateLayer was discarding layer edits if the
new and old layers only differed in their Who object.  The bug
was introduced in the course of removing the layer editor window
and the Display class.

Changed DisplayMethod.inequivalent() to DisplayMethod.equivalent(),
and improved the comments on it and on DisplayMethod.__eq__().
inequivalent() was only used in incorporateLayerSet(), and the extra
negation was unnecessary.