Changed ChooserWidget so that it can choose from any list of objects, not just strings.
The objects have to be convertible to strings with str(), and the
results should be unique.

ChooserWidget.set_state(int) is no longer allowed, unless the list of
objects contains the int.  I think this was only used in one place,
in whowidget.py, and is easily done differently there.

This change was required because the new gtk3 ChooserWidget couldn't
handle the AutomaticName class used in ScheduledOutputs.  The old
gtk+2 widget didn't have a problem because it used an object that
converted objects to strings silently (I think).