Fixed a bug that was affecting field initializers.
If a mesh had conflicting boundary conditions then all of its
subproblems were deemed to be unsolvable, which meant that
SubProblemContext.solver_precompute wouldn't call precomputeMaterials,
which meant that the presence of second order time equations wasn't
detected, which meant that the first order time derivative fields
weren't listed in the initializers pane of the solver page.  Now
the subproblems don't consider boundary conditions when they decide
if they're solvable or not, but they are considered when deciding
if a mesh is solvable.