Switchboard stack tracking now works across all threads, not just ones set up by the switchboard.
All that was necessary was for the routines that create subthreads
or install callbacks on the main thread to wrap the callbacks in
the wrapper that the switchboard was already using.
*However*, doing that broke the exit code, because the wrapper was
trying to get the thread state after threading was turned off.
So I copied the atShutDown mechanism from OOF3D, which allows
other parts of the program to register routines to be run before
the rest of the program quits.  This allows stack tracking to be
turned off in time.  *However*, atShutDown couldn't be imported
by switchboard.spy or quit.py without causing an import loop,
so I put it in its own file.