Working on reducing redundant checkpoints in gui logs.
Moved most of the calls to autoenableBCs so that they're
inside the preceding Who.reserve/Who.cancel_reservation block,
and don't need to make the reservation themselves.  This eliminates
a bunch of gui sensitization calls and their checkpoints, although
not as much as I had hoped.

autoenableBCs used to send a switchboard signal, which now has to
be done by the calling routine so that it can be done outside the
reservation block.  The switchboard notification had a comment about
it being hacky, since it passed dummy arguments.  Now that it uses
the dummy arguments more often than it doesn't, I gave the callback
function (BoundaryCondPage.newBCs) default values so the dummies
aren't needed.

SolverPage.sensitize() can be suppressed until all updates are
complete.