Fixed pdf_compare.  It has to be able to ignore the /Producer stamp in the file as well as the time stamp.
When Cairo writes a pdf file it includes a /Producer string that
contains the Cairo version number.  That number fortunately is
different on the two systems that I'm using, so I caught the error.
The new version of pdf_compare uses regular expressions to search
the data for the time stamp and producer stamp, and then compares
the bytes between the stamps (without copying the strings).  The
strings are searched multiple times (once for each regex and once
to compare what's left) so this version is slower, but more correct.