Changed the RestrictedStringParameter so that it uses a regular expression instead of a list of excluded characters.
Also changed the RestrictedAutomaticNameParameter.

The old method just made sure the names didn't contain a ':' which
would confuse the LabelTrees, but actually for Property names
that's not sufficient.  Property names have to be legal python
variable names.   The former "exclude" argument, listing illegal
characters, is now a "pattern" argument, containing a regular
expression that the input must match.