Fixed a gfxwindow mouse event handling bug.
It is possible for the graphics window to get a mouse-up event without
a corresponding mouse-down event, if the user clicks on a window that
doesn't have focus.  If the mouse-down event gives the window focus
but is not passed along for the window to process, the window will only
see the ensuing mouse-up event.  This is a problem if the window's
mouse handler is expecting to see both events.  With this change
mouse-up events are ignored at the gfxwindow level, before calling
the mouse handler, if there has been no mouse-down.

This change broke some gui tests that had recorded only the mouse-up
events and had acted on them (because the current mouse handler
didn't need a mouse-down event).  On replay, those mouse-up events
are now ignored.  I just added mouse-down events to the affected
log files, so that the up events will no longer be ignored.  In the
future, in those situations, the spurious mouse-ups will be ignored
both during recording or replaying, so there should be no problem
(other than a few extra lines in the log files).