Fixed problems in the ValueOutputWidget.
The hiding and showing of the ScalarOutputWidget and AggregateOutputWidget
wasn't working in gtk3, although there were no significant changes
from gtk2.  The second time the widget was used, both subwidgets
would appear.  I'm not sure why it wasn't working, but now it uses
a GtkStack to switch modes, which is cleaner anyway.