Changed the way that separators are handled in Choosers.
The old way broke the gui tests because the proxy string for
separators was not available in the test scripts, which know
the Gtk objects but not the Chooser object.  Now the separators
are inserted into the popup menu but aren't in the GtkStack,
which is what generics.chooserCheck queries.