Bug fixes for the Canvas Info button in the Viewer toolbox.
The backing layer for the canvas wasn't being initialized properly
on an empty canvas.  This only mattered if you wanted to ask an
empty canvas how big it is, which is a silly question because
the canvas will resize itself as soon as something is drawn.  But
fixing the bug prevents a seg fault if the question is asked.

Changes in gfxwindow.py and viewertoolboxGUI.py fix some
lines leftover from gtk+2 and libgnomecanvas.