Changed the way modifier keys are logged on mouse button press and release events.
The "Repeat" button in the pixel selection toolbox is sensitive to
the state of the shift and control keys, but the keys were being
ignored when replaying a gtklogger file.  Using replay.wevent (which
uses GtkWidget.event() and the documentation says is bad) instead
of replay.event (which uses Gtk.main_do_event, which the documentation
says is preferred) seems to fix the problem.

I'm not sure what the consequences of this will be.