Fixed problems with the misorientation gui in the pixel info toolbox.
The widgets for choosing the lattice symmetry when computing
misorientations in the pixel info toolbox was badly constructed.
If the ConvertibleRegisteredClassFactory was set to SpaceGroup,
it used a LabelledSlider to choose the group number.  Typing in
the GtkEntry or moving the slider invoked a menu command, which
resulted in the widget being rebuilt, and losing focus.  That meant
that only one slider increment or one keystroke would be processed.

Now the toolbox displays the current lattice symmetry, and there's
a "Set" button that allows it to be changed via a dialog box.