Changed the way gui logging works for OOFCanvas events.
gtklogger works by wrapping all GtkWidget callbacks with a function
that logs an event before calling the OOF callback.  But because
the OOFCanvas is not a GtkWidget, this scheme didn't work for it.
So we were making two connections to the GtkLayout that lives inside
the OOFCanvas -- one for logging and one for calling the GfxWindow's
mouse callback.  This meant that we had no control over which ran
first.  If the mouse callback put checkpoints in the log, and if
the callbacks ran in the wrong order, those checkpoints could appear
before the mouse clicks, and the log would hang when replayed.

Now the logging is done by the GfxWindow, which bypasses most of
the gtklogger machinery.  It writes directly to the log file using
gtklogger.writeLine().  This is actually much simpler, and doesn't
require a separate CanvasLogger class anymore.

Added some comments in gtklogger/core.py.

Changed _writeline() in gtklogger/loggers.py to writeLine() so that
it can be called by other parts of the code.

Removed debugging lines from movenodeGUI.py.