Added an Incomplete Cholesky preconditioner, and used it in all tests that used to use ILU with CG.
Also made the names of the preconditioners consistent in the GUI.
Removed ILU from the GUI, since it's the same as ILUT.  It's still
allowed in scripts.

IC is now listed first, before ILUT, in the GUI, so that inexperienced
users might use it by default if they stray into Advanced Solver mode.
Conjugate Gradient is not guaranteed to work with ILUT.

This commit breaks minor things in the GUI tests, but I'm not going
to fix them in this branch.  I'll merge this branch with the gtk3
upgrade branch and fix them there.