Trying to build sparse matrices more efficiently, to reduce memory usage.
Instead of passing a single list of (i,j,val) triplets to
Eigen::SparseMatrix::setFromTriplets, where the list can contain
multiple triplets with the same (i,j), this version adds the values
for the duplicated (i,j)s, tells Eigen the number of non-zero entries
in each column, and inserts the unique triplets directly into the
SparseMatrix.

The hope is that this should reduce memory fragmentation when the
matrices are large.  The list of triplets is really a list of lists,
where each sublist contains the triplets for a single column of the
matrix.  These lists are not going to be very large.  The list of
lists is large, but never needs to be reallocated.   Keeping a
separate list for each column makes it easy to compress the lists
(combining entries that have the same i,j) and to tell Eigen how
much space to reserve, so that it doesn't need to do any reallocation
when building the matrix.

The ChunkyVectors that LinearizedSystem was using for the list
of Triplets have been removed.

There are some optimizations that can be made.  In particular there's
no need to use Eigen::Triplet anymore, so we can use a class that
allows the values to be added in place.