Fixed some bugs in the new matrix construction method.
The temporary matrices for the Triplets weren't being reconstructed
when a LinearizedSystem object was reused by a nonlinear solver.

Now LinearizedSystem::consolidate takes a bool argument that indicates
whether or not it will be reused.  If it will be, the storage is
cleared but not deallocated.

Also deleted some memusage calls that weren't in useful locations.