Reduced the memory required in sparse matrix construction a bit.
Introducing a new class, Doublet, to replace Eigen::Triplet<double>
in matrix construction.  Doublet doesn't store the column index,
since when the Doublets are created in LinearizeSystem they're
stored in separate arrays for each column.   Also, Doublet::value()
returns a non-const reference so that Doublets that share a row can
be consolidated easily.

All the relevant variables and methods that referred to "triplet"
or "tri" have been changed.