Changed the way that selection changes are logged by the GfxLabelTree.
Logging the GtkTreeSelection "changed" signal is problematic because
the signal is sometimes emitted when nothing has changed, and is
emitted in different situations when recording and replaying, making
checkpoints fail in its callback function.

The GfxLabelTree therefore does *not* log the GtkTreeSelection
"changed" signal.  It connects to the signal directly (ie, not via
gtklogger) and inserts its own code in the gui log file when it
determines that something has changed.  This method is described
in the gtklogger README file, which also includes a TODO about maybe
making the process generic so that it can be easily applied to all
GtkTreeViews.

To use the new code, the GfxLabelTree constructor's "name" argument
is now required, so it's no longer an optional keyword argument.