Added a way to set the title for an automatically generated OOFMenu dialog.
When the layer editor window was replaced by a simple menu command,
the "New" menu item in the graphics window's Layer menu didn't need
a custom gui callback.  But the default gui callback used the menu
item name as the title for the automatically generated dialog box,
and "New" was an uninformative name.  Now OOFMenuItems can be given
an optional gui_title argument when constructed, specifying a string
to be used instead of the menu item name as the title for the dialog
box.

This change broke a large number of gui tests, which will be fixed
in the next commit.  The gtklogger name of the dialog box and
checkpoints based on the name all changed.