Changes to a couple of gui tests caused by merging the efficient-sparse branch.
With the new version of Eigen in the efficient-sparse branch, the
default preconditioner has changed.

This version passes the regression and gui tests on Mac using quartz,
but fails the gui test using X11, which is what it did before the
merge too.  I'm committing it now anyway, so that I can test it on Linux.