Added a test for converting orientations from one representation to another.
The checkOrientationWidget() function in generics.py  was defective
and is now better.   It was computing relative tolerances incorrectly
when comparing numbers, and comparing any value to zero would fail,
even if within the tolerance.  I fixed it, and also changed the
way that other routines were doing it, because they could in principle
have the same problem.