Removed a TODO about GtkLoggerTopFailure and improved its comment.
The TODO was about checking to see if we still need to repeat
gui log lines that raise GtkLoggerTopFailure when replayed.
It appears that the current OOF2 test suite does work without
repeating any lines, meaning that the checkpoints are sufficient
to prevent GtkLoggerTopFailures.  However, there's also no need
to remove the code that repeats lines until the failure goes away,
just in case there are cases that checkpoints don't resolve.